refactor(api): gitignore, model naming
