+++ packages/api/src/modules/note/note.entity.ts
@@ -1,6 +1,6 @@
 import { Field, ObjectType } from '@nestjs/graphql'
 
-import { User } from '../user/user.model'
+import { User } from '../user/user.entity'
 
 @ObjectType()
 export class Note {
