+++ packages/api/src/modules/user/dtos/user.create.input.ts
@@ -1,5 +1,5 @@
 import { Field, InputType } from '@nestjs/graphql'
-import { User } from '../user.model'
+import { User } from '../user.entity'
 
 @InputType()
 export class UserCreateInput implements Partial<User> {
